#821 [REFACTOR] shared/component/modal 분류 완료
