--- src/components/Modal/ConfirmModal/index.js
@@ -1,1 +1,0 @@
-export { default as ConfirmModal } from './ConfirmModal';
