--- src/components/Modal/index.js
@@ -1,5 +1,0 @@
-export * from './ConfirmModal';
-export { default as Options } from './Options';
-export { default as OptionModal } from './OptionModal';
-export { default as DeleteModal } from './DeleteModal';
-export { default as TextAreaModal } from './TextAreaModal';
