+++ src/components/ReviewDownload/ReviewDownload.jsx
@@ -6,8 +6,7 @@ import { getExamReview } from '@/apis';
 
 import { useToast } from '@/hooks';
 
-import { FetchLoadingOverlay } from '@/shared/component';
-import { DeleteModal } from '@/components/Modal';
+import { DeleteModal, FetchLoadingOverlay } from '@/shared/component';
 import { Icon } from '@/components/Icon';
 
 import { LOADING_MESSAGE, QUERY_KEY, TOAST } from '@/constants';
