+++ src/components/index.js
@@ -16,7 +16,6 @@ export * from './ListHeader';
 export * from './MainPageListItem';
 export * from './Margin';
 export * from './MenuIcon';
-export * from './Modal';
 export * from './Portal';
 export * from './NoticeBar';
 export * from './PostBar';
