+++ src/pages/ExamReviewPage/ExamReviewPage.jsx
@@ -12,13 +12,13 @@ import { useAuth, useScrap, useToast } from '@/hooks';
 
 import {
   BackAppBar,
+  DeleteModal,
   FetchLoading,
   FetchLoadingOverlay,
+  OptionModal,
 } from '@/shared/component';
 import {
   CommentsSuspense,
-  DeleteModal,
-  OptionModal,
   Icon,
   InputBar,
   ReviewContentItem,
